+++ android/expoview/src/main/java/versioned/host/exp/exponent/ExponentPackage.java
@@ -41,6 +41,7 @@ import versioned.host.exp.exponent.modules.api.URLHandlerModule;
 import versioned.host.exp.exponent.modules.api.UpdatesModule;
 import versioned.host.exp.exponent.modules.api.cognito.RNAWSCognitoModule;
 import versioned.host.exp.exponent.modules.api.components.datetimepicker.RNDateTimePickerPackage;
+import versioned.host.exp.exponent.modules.api.components.maskedview.RNCMaskedViewPackage;
 import versioned.host.exp.exponent.modules.api.components.gesturehandler.react.RNGestureHandlerModule;
 import versioned.host.exp.exponent.modules.api.components.gesturehandler.react.RNGestureHandlerPackage;
 import versioned.host.exp.exponent.modules.api.components.lottie.LottiePackage;
@@ -217,7 +218,8 @@ public class ExponentPackage implements ReactPackage {
         new RNCWebViewPackage(),
         new SafeAreaContextPackage(),
         new RNSharedElementPackage(),
-        new RNDateTimePickerPackage()
+        new RNDateTimePickerPackage(),
+        new RNCMaskedViewPackage()
     ));
 
     viewManagers.addAll(mModuleRegistryAdapter.createViewManagers(reactContext));