+++ apps/native-component-list/src/navigation/ReactNativeCoreStackNavigator.ts
@@ -1,5 +1,3 @@
-import BasicMaskExample from '../screens/BasicMaskScreen';
-import GLMaskExample from '../screens/MaskGLScreen';
 import ReactNativeCore from '../screens/ReactNativeCore/ReactNativeCoreScreen';
 import createStackNavigator from './createStackNavigator';
 import StackConfig from './StackConfig';
@@ -7,8 +5,6 @@ import StackConfig from './StackConfig';
 const ReactNativeCoreStackNavigator = createStackNavigator(
   {
     ReactNativeCore,
-    BasicMaskExample,
-    GLMaskExample,
   },
   StackConfig
 );